+++ app/Config/Filters.php
@@ -33,7 +33,7 @@ class Filters extends BaseConfig
 	public $globals = [
 		'before' => [
 			// 'honeypot',
-			'csrf',
+			'csrf' => ['except' => ['api/v1/*']]
 		],
 		'after'  => [
 			'toolbar',