+++ app/Models/UserModel.php
@@ -8,9 +8,14 @@ class UserModel extends MythModel
     protected $returnType = User::class;
  
     protected $allowedFields = [
-        'email', 'username', 'password_hash', 'reset_hash', 'reset_at', 'reset_expires', 'activate_hash',
+        'email', 'password_hash', 'reset_hash', 'reset_at', 'reset_expires', 'activate_hash',
         'status', 'status_message', 'active', 'force_pass_reset', 'permissions', 'deleted_at',
         'name', 'lastname',
+    ];
+
+    protected $validationRules = [
+        'email'         => 'required|valid_email|is_unique[users.email,id,{id}]',
+        'password_hash' => 'required',
     ];
  
 }